app/requirements.txt
click==7.1.2
fastapi==0.63.0
mangum
h11==0.12.0
joblib==1.0.1
numpy==1.20.2
Pillow==8.2.0
pydantic==1.8.1
starlette==0.13.6
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.7.4.3
uvicorn==0.13.4
